Create LayfarersCronicle.js
This is the completed script as of 01/16/2024. This contains
the following ancestries + traits related with them:
- Birdfolk
- Catfolk
- Elf
- Gnome
- Human
- Lizardfolk
- Orc
- Ratfolk

Each has the option for a 'mixed lineage' as well as optional origin
selections (if available.)

Traits are selected via the 'Feats' section on the third page. This was
done because there is not a great way to contain multiple options for
ancestry creation in the `RacialVariant` list. The overflow page
should be enabled to access actual feats past level 4.